Fixed 64-bit GCC warnings in MESS-specific code (src/emu still has some issues).
git-svn-id: svn://messdev.no-ip.org/mess@750 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/tools/messtest/core.c
@@ -106,7 +106,7 @@ void messtest_get_data(xml_data_node *node, mess_pile *pile)
 	blobparse_state_t blobstate = BLOBSTATE_INITIAL;
 	const char *s;
 	size_t len;
-	int multiple = 0;
+	size_t multiple = 0;
 
 	s = node->value ? node->value : "";
 	len = strlen(s);
